% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{Geographical variation in bill size and tarsus length across bird species}
\format{
A data frame (tibble) with 215 rows and 9 variables:
\describe{
\item{family}{string: family name}
\item{genus}{string: genus name}
\item{species}{string: species name}
\item{mass}{number: mass in grams}
\item{bill}{number: length in mm}
\item{tarsus}{number: length in mm}
\item{latitude}{number: midpoint of range from equator in degrees}
\item{altitude}{number: midpoint of range in meters}
\item{temperature}{number: minimum temperature at midpoint of range in degrees C}
}
}
\source{
Symonds, Matthew R. E. and Glenn J. Tattersall. 2010. Geographical Variation
in Bill Size across Bird Species Provides Evidence for Allen’s Rule. Am Nat 176:188-197.
\url{https://doi.org/10.1086/653666} Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.1421}
}
\usage{
birds
}
\description{
These data are from a study evaluating Allen's rule, which is the proposition
that warm-blooded animals have shorter extremities in colder climates to aid
in heat retention. In this study, the authors evaluated this proposition across
bird species for bill and tarsus (leg) length.
}
\keyword{datasets}
