% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloodpressure}
\alias{bloodpressure}
\title{Blood pressure and covariates data for twenty hypertensive individuals}
\format{
A data frame (tibble) with 20 rows and 7 variables:
\describe{
\item{bp}{number: blood pressure in mm Hg}
\item{age}{number: age in years}
\item{weight}{number: weight in kg}
\item{bsa}{number: body surface area in sq m}
\item{duration}{number: duration of hypertension in years}
\item{pulse}{number: basal pulse in beats per minute}
\item{stress}{number: stress index}
}
}
\source{
I don't know that this dataset ultimately originates here,
but I found it on the website for STAT 501 at Penn State
\url{https://online.stat.psu.edu/stat501/}
}
\usage{
bloodpressure
}
\description{
This seems to be a commonly used exemplar dataset.
}
\keyword{datasets}
