% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boas_h2}
\alias{boas_h2}
\title{A subset of Boas' immigrant study data suitable for calculating heritability}
\format{
A data frame with 162 rows and 22 variables:
\describe{
\item{fam_id2}{character, unique family identifier}
\item{sex}{factor, "m" or "f"}
\item{age}{numeric, age in years}
\item{hd_lng}{numeric, head length in mm}
\item{hd_wdt}{numeric, maximum head width in mm}
\item{bizyg}{numeric, bizygomatic breadth in mm}
\item{stature}{numeric, height in cm}
\item{cephalic}{numeric, cephalic index}
\item{wfi}{numeric, facial index}
\item{usborn}{factor, "foreign-born" or "us-born"}
\item{m_hd_lng}{numeric, mother's head length in mm}
\item{m_hd_wdt}{numeric, mother's maximium head width in mm}
\item{m_bizyg}{numeric, mother's bizygomatic breadth in mm}
\item{m_stat}{numeric, mother's height in cm}
\item{m_ceph}{numeric, mother's cephalic index}
\item{m_wfi}{numeric, mother's facial index}
\item{f_hd_lng}{numeric, father's head length in mm}
\item{f_hd_wdt}{numeric, father's maximum head width in mm}
\item{f_bizyg}{numeric, father's bizygomatic breadth in mm}
\item{f_stat}{numeric, father's height in cm}
\item{f_ceph}{numeric, father's cephalic index}
\item{f_wfi}{numeric, father's facial index}
}
}
\source{
Boas, Franz. Changes in Bodily Form of Descendants of Immigrants.
New York: Columbia University Press, 1912.

Boas, Franz. “Changes in Bodily Form of Descendants of Immigrants.”
Washington, DC: United States Senate, 1910.

Boas, Franz. “Changes in the Bodily Form of Descendants of Immigrants.”
American Anthropologist 14, no. 3 (1912): 530–562.

Gravlee, Clarence C., H. Russell Bernard, and William R. Leonard.
“Heredity, Environment, and Cranial Form: A Reanalysis of Boas’s Immigrant
Data.” American Anthropologist 105, no. 1 (March 2003): 125–38.
\url{https://doi.org/10.1525/aa.2003.105.1.125}

Gravlee, Clarence C., H. Russell Bernard, and William R. Leonard. “Boas’s
Changes in Bodily Form: The Immigrant Study, Cranial Plasticity, and
Boas’s Physical Anthropology.” American Anthropologist 105, no. 2
(June 2003): 326–32.
\url{https://doi.org/10.1525/aa.2003.105.2.326}

Raw data deposited in openICPSR
\url{https://www.openicpsr.org/openicpsr/project/112086/version/V2/view}
}
\usage{
boas_h2
}
\description{
This subset of Boas' immigrant study data includes only adult children
where stature data are available for both parents from what Boas labelled
the "Hebrew" immigrant group sample. Between 1880 and 1924, approximately
2.5 million Eastern European Ashkenazim immigrated to the United States,
many settling in New York City. Boas' "Hebrew" sample is
undoubtedly primarily of this population.
}
\details{
This subset is purely for
demonstration purposes - researchers should access the full dataset
freely accessible at the link below.
}
\keyword{datasets}
