% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boas_stature}
\alias{boas_stature}
\title{A subset of stature data from Boas' immigrant study data}
\format{
A data frame with 12,641 rows and 7 variables:
\describe{
\item{id}{numeric, unique identifier}
\item{immgroup}{factor, one of "sicilian", "central_italian", "bohemian",
"hungarian_slovak", "polish", "scotch", or "hebrew"}
\item{sex}{factor, "m" or "f"}
\item{imm_yr}{numeric, immigration year; US-born individuals are coded -999}
\item{age}{numeric, age in years}
\item{stature}{numeric, height in cm}
\item{usborn}{factor, "foreign-born" or "us-born"}
}
}
\source{
Boas, Franz. Changes in Bodily Form of Descendants of Immigrants.
New York: Columbia University Press, 1912.

Boas, Franz. “Changes in Bodily Form of Descendants of Immigrants.”
Washington, DC: United States Senate, 1910.

Boas, Franz. “Changes in the Bodily Form of Descendants of Immigrants.”
American Anthropologist 14, no. 3 (1912): 530–562.

Gravlee, Clarence C., H. Russell Bernard, and William R. Leonard.
“Heredity, Environment, and Cranial Form: A Reanalysis of Boas’s Immigrant
Data.” American Anthropologist 105, no. 1 (March 2003): 125–38.
\url{https://doi.org/10.1525/aa.2003.105.1.125}

Gravlee, Clarence C., H. Russell Bernard, and William R. Leonard. “Boas’s
Changes in Bodily Form: The Immigrant Study, Cranial Plasticity, and
Boas’s Physical Anthropology.” American Anthropologist 105, no. 2
(June 2003): 326–32.
\url{https://doi.org/10.1525/aa.2003.105.2.326}

Raw data deposited in openICPSR
\url{https://www.openicpsr.org/openicpsr/project/112086/version/V2/view}
}
\usage{
boas_stature
}
\description{
This subset of Boas' immigrant study data includes all individuals with
data on both age and stature.
}
\details{
This subset is purely for
demonstration purposes - researchers should access the full dataset
freely accessible at the link below.
}
\keyword{datasets}
