% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora_info}
\alias{carnivora_info}
\title{Body mass and life history trait infomation for carnivore species}
\format{
A data frame (tibble) with 26 rows and 11 variables:
\describe{
\item{family}{character, taxonomic family}
\item{species}{character, binomial species name}
\item{mass_male}{numeric, mass in kg}
\item{mass_female}{numeric, mass in kg}
\item{mass_mean}{numeric, mass in kg}
\item{polygyny}{factor, polygynous or non_polygynous}
\item{parental_care}{factor, female_only or female_with_helpers}
\item{diet}{factor, carnivorous or non_carnivorous}
\item{locomotor_zone}{factor, terrestrial or terrestrial_arboreal}
\item{social_group_size}{factor, solitary or social}
\item{foraging_group_size}{factor, forage_along or forage_in_groups}
}
}
\source{
Morris, J.S. and Carrier, D.R. (2016), Sexual selection on skeletal
shape in Carnivora. Evolution, 70: 767-780.
\url{https://doi.org/10.1111/evo.12904}
Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.86523}
}
\usage{
carnivora_info
}
\description{
A set of body mass and life history trait information for carnivores from a study
examining sexual selection in carnivores. See also \code{\link{carnivora_indices}}.
}
\keyword{datasets}
