% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citytemps}
\alias{citytemps}
\title{Average January temperature for US and European cities}
\format{
A data frame (tibble) with 83 rows and 4 variables:
\describe{
\item{city}{a string with the city name}
\item{temperature}{a number with the average daily high temperature in January in degrees C}
\item{latitude}{a number with the city's latitude in degrees}
\item{region}{a factor - US or Europe}
}
}
\source{
I've lost track of the original source - some online almanac
of some sort.
}
\usage{
citytemps
}
\description{
A dataset somewhat selectively chosen to illustrate Simpson's Paradox.
}
\keyword{datasets}
