% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy}
\alias{energy}
\title{Simulated energy expenditure and travel data}
\format{
A data frame (tibble) with 90 rows and 3 variables:
\describe{
\item{expend}{number: energy expenditure in kilocalories}
\item{travel}{number: average daily path in meters}
\item{group}{factor: female, pregnant female, or male}
}
}
\source{
Data simulated by Ryan Raaum.
}
\usage{
energy
}
\description{
This data table presents energy expenditure and average daily
path length for males, females, and pregnant females of a
hypothetical primate species.
}
\keyword{datasets}
