% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flyrules}
\alias{flyrules}
\title{Subset of Flying Etiquette Survey Data}
\format{
A data frame (tibble) with 846 rows and 4 variables:
\describe{
\item{do_you_recline}{Do you ever recline your seat when you fly?}
\item{rude_to_recline}{Is it rude to recline your seat on a plane?}
\item{has_child_under_18}{Do you have any children under 18?}
\item{rude_to_bring_baby}{In general, is it rude to bring a baby on a plane?}
}
}
\source{
Subsetted from the \code{fly} dataset in the \code{ggmosaic} package,
ultimately originating from \url{https://github.com/fivethirtyeight/data/tree/master/flying-etiquette-survey}
}
\usage{
flyrules
}
\description{
This is a subset of data from the results of a SurveyMonkey survey
commissioned by FiveThirtyEight for the story 41 Percent of Fliers
Say It’s Rude To Recline Your Airplane Seat.
}
\keyword{datasets}
