% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gfortis}
\alias{gfortis}
\title{\emph{Geospiza fortis} beak depth data}
\format{
A data frame with 413 rows and 8 variables:
\describe{
\item{brood_id}{factor: identifier for each brood of offspring}
\item{species}{string: taxonomic species}
\item{mid_offspring_bd}{number: average beak depth across all offspring in brood in mm}
\item{male_id}{factor: identifier for presumed male parent}
\item{male_bd}{number: beak depth of presumed male parent in mm}
\item{female_id}{number: identifier for female parent}
\item{female_bd}{number: beak depth of female parent in mm}
\item{mid_parent_bd}{number: average of male and female parent beak depths in mm}
}
}
\source{
Grant, P., & Grant, B. (2014). 40 Years of Evolution:
Darwin's Finches on Daphne Major Island. Princeton; Oxford:
Princeton University Press. \url{https://doi.org/10.2307/j.ctt5hhncc}
Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.g6g3h}
}
\usage{
gfortis
}
\description{
These data are the basis for the estimation of beak depth heritability
in Grant & Grant 2014 using the parent-offspring regression approach.
}
\details{
Changes from the data as provided by the Grant's are:
\itemize{
\item variable names changed to be more consistent and R-friendly
\item the brood identifier in the original is - inexplicably - the average
value of the offspring band numbers. Here they are relabeled with sequential
BXXX labels.
}
}
\keyword{datasets}
