% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gfortis_ped}
\alias{gfortis_ped}
\title{\emph{Geospiza fortis} pedigree data}
\format{
A data frame with 979 rows and 3 variables:
\describe{
\item{id}{factor: identifier for offspring}
\item{male_id}{factor: identifier for male parent}
\item{female_id}{factor: identifier for female parent}
}
}
\source{
Grant, P., & Grant, B. (2014). 40 Years of Evolution:
Darwin's Finches on Daphne Major Island. Princeton; Oxford:
Princeton University Press. \url{https://doi.org/10.2307/j.ctt5hhncc}
Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.g6g3h}
}
\usage{
gfortis_ped
}
\description{
This pedigree was derived from the \code{\link{gfortis}} data for the
purposes of providing an example of estimating heritability using the
animal model. \strong{NOTE:} As the \code{\link{gfortis}} data do not have
individual entries for all offspring, but rather brood averages, the
heritability estimate from the animal model will be inflated due to the
loss of residual variance in the brood-averaging.
}
\seealso{
\code{\link{gfortis}} \code{\link{gfortis_bd}}
}
\keyword{datasets}
