% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grooming}
\alias{grooming}
\title{Simulated grooming duration data}
\format{
A data frame (tibble) with 100 rows and 5 variables:
\describe{
\item{focal}{factor: identifier for one of 15 adult females}
\item{relrank}{factor: relative rank of animal being groomed to focal individual, higher or lower}
\item{groomtime}{number: duration of grooming in seconds}
\item{age}{number: age of focal individual in years}
\item{matriline}{factor: identifier for one of 4 matrilines}
}
}
\source{
Data simulated by Ryan Raaum.
}
\usage{
grooming
}
\description{
Simulated dataset of grooming duration within a group of
female-bonded primates
}
\keyword{datasets}
