% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jawtree}
\alias{jawtree}
\title{A phylogenetic tree of a subset of primates}
\format{
A tree in the \code{ape} package's 'phylo' object
}
\source{
Meloro, C., Cáceres, N.C., Carotenuto, F., Sponchiado, J., Melo, G.L.,
Passaro, F. and Raia, P. (2015), Chewing on the trees: Constraints and
adaptation in the evolution of the primate mandible. Evolution, 69: 1690-1700.
\url{https://doi.org/10.1111/evo.12694}
Raw data deposited in figshare \url{https://doi.org/10.5061/dryad.8jq98}
}
\usage{
jawtree
}
\description{
This is a tree to be used in a phylogenetic PCA along with
\code{\link{jawshape}}.
}
\seealso{
\code{\link{jawshape}}
}
\keyword{datasets}
