% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lilpantheria}
\alias{lilpantheria}
\title{This is a subset of the PanTHERIA dataset}
\format{
A data frame (tibble) with 157 rows and 8 variables:
\describe{
\item{order}{string: taxonomic order}
\item{family}{string: taxonomic family}
\item{genus}{string: taxonomic genus}
\item{species}{string: taxonomic species}
\item{length}{number: head+body length in mm}
\item{mass}{number: mass in g}
\item{afb}{number: age at first birth in days}
\item{interbirth}{number: interbirth interval in days}
}
}
\source{
Jones, K.E., Bielby, J., Cardillo, M., Fritz, S.A.,
O'Dell, J., Orme, C.D.L., Safi, K., Sechrest, W., Boakes, E.H.,
Carbone, C., Connolly, C., Cutts, M.J., Foster, J.K.,
Grenyer, R., Habib, M., Plaster, C.A., Price, S.A., Rigby, E.A.,
Rist, J., Teacher, A., Bininda-Emonds, O.R.P., Gittleman, J.L.,
Mace, G.M. and Purvis, A. (2009), PanTHERIA: a species-level
database of life history, ecology, and geography of extant
and recently extinct mammals. Ecology, 90: 2648-2648.
\url{https://doi.org/10.1890/08-1494.1}
Raw data deposited in figshare \url{https://doi.org/10.6084/m9.figshare.c.3301274.v1}
}
\usage{
lilpantheria
}
\description{
Body length, mass, age at first birth, and interbirth interval for
all primates, rodents, and artiodactyls in the PanTHERIA compilation
that have no missing data for these variables.
}
\keyword{datasets}
