% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Mass and basal metabolic rate data for mammals}
\format{
A data frame (tibble) with 696 rows and 9 variables:
\describe{
\item{order}{string: taxonomic order}
\item{family}{string: taxonomic family}
\item{genus}{string: taxonomic genus}
\item{species}{string: taxonomic species}
\item{bodytemp}{number: body temperature in degrees C}
\item{mass}{number: mass in g}
\item{mass_sd}{number: within-species mass standard deviation}
\item{bmr}{number: basal metabolic rate in mL O2 / hr}
\item{bmr_sd}{number: within-species bmr standard deviation}
}
}
\source{
Sieg, Annette E., Michael P. O’Connor, James N. McNair, Bruce W. Grant,
Salvatore J. Agosta, and Arthur E. Dunham. 2009. Mammalian Metabolic Allometry:
Do Intraspecific Variation, Phylogeny, and Regression Models Matter?
Am Nat 174:720-733. \url{https://doi.org/10.1086/606023}
Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.712}
}
\usage{
mammals
}
\description{
These data are from a study evaluating the hypothesis that there is a
universal metabolic rate–body mass scaling relationship.
}
\keyword{datasets}
