% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pefr}
\alias{pefr}
\title{Example data for matched pair analysis}
\format{
A data frame (tibble) with 9 rows and 3 variables:
\describe{
\item{before}{number: pefr in liters/minute before walking}
\item{after}{number: pefr in liters/minute after walking}
\item{subject}{factor: subject identifier}
}
}
\source{
Online in many places - seems to be a commonly used exemplar.
Not sure what the original source is.
}
\usage{
pefr
}
\description{
Comparison of peak expiratory flow rate (PEFR) before and after a walk
on a cold winter's day for a random sample of 9 asthmatics. The peak
expiratory flow rate measures how fast a person can breathe out (exhale)
air. It is one of many tests that measures how well the lungs are working.
Measured in liters/minute.
}
\keyword{datasets}
