% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primatetree}
\alias{primatetree}
\title{This is the 10k Trees Project consensus primate tree}
\format{
A tree in the \code{ape} package's 'phylo' object
}
\source{
\url{https://10ktrees.nunn-lab.org}
}
\usage{
primatetree
}
\description{
This is a tree to be used in a phylogenetic regression along with
\code{\link{primatedata}}.
}
\seealso{
\code{\link{primatedata}}
}
\keyword{datasets}
