% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvals}
\alias{pvals}
\title{Simulated p-values for multiple testing demonstration}
\format{
A data frame (tibble) with 1000 rows and 2 variables:
\describe{
\item{null}{logical: whether or not the null hypothesis is true}
\item{p}{number: p-value}
}
}
\source{
Data simulated by Ryan Raaum.
}
\usage{
pvals
}
\description{
Simulated dataset of p-values where whether or not the null hypothesis
is true is known.
}
\keyword{datasets}
