% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teedata}
\alias{teedata}
\title{Total energy expenditure across primate species}
\format{
A data frame with 84 rows and 5 variables:
\describe{
\item{species}{string: taxonomic species}
\item{order}{string: taxonomic order}
\item{group}{factor: primate or nonprimate}
\item{mass}{number: mass in kg}
\item{TEE}{number: total energy expenditure in kcalories/day}
}
}
\source{
Herman Pontzer, David A. Raichlen, Adam D. Gordon,
Kara K. Schroepfer-Walker, Brian Hare, Matthew C. O’Neill,
Kathleen M. Muldoon, Holly M. Dunsworth, Brian M. Wood,
Karin Isler, Judith Burkart, Mitchell Irwin, Robert W.
Shumaker, Elizabeth V. Lonsdorf, Stephen R. Ross. 2014.
Primate energy expenditure and life history.
Proceedings of the National Academy of Sciences 111:1433-1437
\url{https://doi.org/10.1073/pnas.1316940111}
}
\usage{
teedata
}
\description{
Data from "Primate energy expenditure and life history". Used
for a phylogenetic regression exercise.
}
\details{
The data are modified from what was released by the authors:
\itemize{
\item called the admixed Eulemur sp. data Eulemur fulvus
\item dropped the captive Lemur catta data
\item dropped the "captive" westerners Homo sapiens data
\item dropped the second Peromyscus maniculatus data point (which
was not used in primate - nonprimate comparisons in Pontzer et al.)
\item added a "group" varible to sort out primates from non-primates
}
}
\seealso{
\code{\link{teetree}}
}
\keyword{datasets}
