% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teetree}
\alias{teetree}
\title{Phylogenetic tree of primates}
\format{
A tree in the \code{ape} package's 'phylo' object
}
\source{
Bininda-Emonds, O., Cardillo, M., Jones, K. et al.
The delayed rise of present-day mammals. Nature 446, 507–512
(2007).\url{https://doi.org/10.1038/nature05634}
}
\usage{
teetree
}
\description{
This is a tree to be used in a phylogenetic regression along with
\code{\link{teedata}}.
}
\details{
The tree from Binida-Emonds et al. 2007 was modified:
\itemize{
\item reduced to taxa shared with the Pontzer et al. 2014 data
\item Proteles_cristatus was renamed to Proteles_cristata
}
}
\seealso{
\code{\link{teedata}}
}
\keyword{datasets}
