% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tusksize}
\alias{tusksize}
\title{Tusk length and circumference of African elephants}
\format{
A data frame (tibble) with 593 rows and 7 variables:
\describe{
\item{time}{factor: time period, 1966-68 or 2005-13}
\item{id}{string: elephant identifier}
\item{sex}{factor: f or m}
\item{age}{number: age in years}
\item{height}{number: height at shoulder in cm}
\item{length}{number: exposed tusk length in cm}
\item{circumference}{number: tusk circumferance at lip line in cm}
}
}
\source{
Chiyo, P.I., Obanda, V. and Korir, D.K. (2015), Illegal tusk
harvest and the decline of tusk size in the African elephant.
Ecol Evol, 5: 5216-5229. \url{https://doi.org/10.1002/ece3.1769}
Raw data deposited in Dryad \url{https://doi.org/10.5061/dryad.h6t7j}
}
\usage{
tusksize
}
\description{
These data are from a study evaluating tusk length and circumference
in African elephants before widespread poaching (1966-68) and
after (2005-13).
}
\keyword{datasets}
